add reloader flag, disable flask logging elsewhere
